reorganize attributes into folders
