+++ Src/Attributes/Attributes/Binding/QueryParamAttribute.cs
@@ -1,0 +1,7 @@
+﻿namespace FastEndpoints;
+
+/// <summary>
+/// properties decorated with this attribute will have a corresponding request parameter added to the swagger schema
+/// </summary>
+[AttributeUsage(AttributeTargets.Property)]
+public class QueryParamAttribute : Attribute;
