+++ Src/Attributes/Attributes/Config/HttpAttribute.cs
@@ -1,0 +1,33 @@
+﻿namespace FastEndpoints;
+
+/// <summary>
+/// base http attribute class
+/// </summary>
+[AttributeUsage(AttributeTargets.Class)]
+public abstract class HttpAttribute : Attribute
+{
+    internal Http Verb { get; set; }
+    internal string[] Routes { get; set; }
+
+    /// <summary>
+    /// constructor
+    /// </summary>
+    /// <param name="verb">verb</param>
+    /// <param name="route">route</param>
+    protected HttpAttribute(Http verb, string route)
+    {
+        Verb = verb;
+        Routes = [route];
+    }
+
+    /// <summary>
+    /// constructor
+    /// </summary>
+    /// <param name="verb">verb</param>
+    /// <param name="routes">routes</param>
+    protected HttpAttribute(Http verb, params string[] routes)
+    {
+        Verb = verb;
+        Routes = routes;
+    }
+}
